Rename casestats.ts to refreshcasestats.ts
