+++ server/functions/src/index.ts
@@ -1,7 +1,7 @@
 import * as admin from "firebase-admin";
 import * as functions from "firebase-functions";
 import { SERVING_REGION } from "./config";
-import * as casestats from "./casestats"; // Is this right?
+import * as refreshcasestats from "./refreshcasestats"; // Is this right?
 import * as who from "./who";
 
 admin.initializeApp();
@@ -22,8 +22,8 @@ interface Client {
   subscribedTopics: string[];
 }
 
-casestats.setDb(db); /// This seems wrong???????????????????????????????????????????????????????????
-export const refreshCaseStats = casestats.refreshCaseStats;
+refreshcasestats.setDb(db); /// This seems wrong???????????????????????????????????????????????????????????
+export const refreshCaseStats = refreshcasestats.refreshCaseStats;
 
 async function getCaseStatsAsync(req: who.GetCaseStatsRequest) {
   let jurisdictions = req.jurisdictions;
