+++ server/functions/src/refreshcasestats.ts
